+++ tranco-sig-sites/scrape.js
@@ -2,8 +2,6 @@ const puppeteerLib = require('puppeteer-extra')
 const stealthPluginLib = require('puppeteer-extra-plugin-stealth')
 const tldtsLib = require('tldts')
 
-const urlsLib = require('./urls')
-
 const __closePage = async (page) => {
   try {
     await page.close()
@@ -26,7 +24,7 @@ const closeAllPages = async (browser) => {
   }
 }
 
-const getPageContentForUrl = async (browser, url, timeoutSecs) => {
+const getPageContentForURL = async (browser, url, timeoutSecs) => {
   const page = await browser.newPage()
   page.setDefaultNavigationTimeout(timeoutSecs * 1000)
 
@@ -34,70 +32,74 @@ const getPageContentForUrl = async (browser, url, timeoutSecs) => {
   try {
     rs = await page.goto(url)
   } catch (e) {
-    console.error('getPageContentForUrl')
-    console.error(e)
-    await __closePage(page)
+    console.error(`getPageContentForURL: url: ${url}`)
+    console.error(`${e.name}: ${e.message}`)
     return null
   }
   if (!rs || rs.ok() === false) {
-    await __closePage(page)
     return false
   }
 
-  const summary = {
-    content: await rs.content(),
-    page,
-    url: rs.url()
-  }
-  return summary
-}
-
-const getSameSitePageUrls = async (page, urlsToIgnore, stripQueryParams) => {
-  const processUrl = urlsLib.process.bind(undefined, stripQueryParams)
-  let pageSite
-  let pageUrl
-  const pageUrlStr = processUrl(page.url())
   try {
-    pageUrl = new URL(pageUrlStr)
-    pageSite = tldtsLib.parse(pageUrl.hostname).domain
+    const summary = {
+      content: await rs.content(),
+      page,
+      url: rs.url()
+    }
+    return summary
   } catch (e) {
-    console.error('getSameSitePageUrls')
-    console.error(e)
+    console.error(`getPageContentForURL: url: ${url}`)
+    console.error(`${e.name}: ${e.message}`)
     return null
   }
+}
 
-  const sameSiteLinkedUrls = new Set()
+const isURLinSet = (urlComparisonFunc, aSet, needleURL) => {
+  for (const aURL of aSet) {
+    if (urlComparisonFunc(aURL, needleURL)) {
+      return true
+    }
+  }
+  return false
+}
+
+const getSameSiteLinkedURLs = async (page, ignoreURLsSet, urlComparisonFunc) => {
+  const shouldIgnoreURL = isURLinSet.bind(undefined, urlComparisonFunc)
+
+  const pageURL = new URL(page.url())
+  const pageSite = tldtsLib.parse(pageURL.hostname).domain
+
+  const sameSiteURLs = new Set()
   const anchorTags = await page.$$('a[href]')
   for (const anchorTag of anchorTags) {
     try {
       const hrefHandle = await anchorTag.getProperty('href')
       const hrefValue = await hrefHandle.jsonValue()
-      const anchorUrl = new URL(hrefValue, pageUrl)
-      const anchorSite = tldtsLib.parse(anchorUrl.hostname).domain
-      const processedAnchorUrlStr = processUrl(anchorUrl.toString())
+      const anchorURL = new URL(hrefValue, pageURL)
+      const anchorSite = tldtsLib.parse(anchorURL.hostname).domain
 
-      if (urlsToIgnore.has(hrefValue) || urlsToIgnore.has(processedAnchorUrlStr)) {
+      if (shouldIgnoreURL(ignoreURLsSet, anchorURL)) {
         continue
       }
-      if (processedAnchorUrlStr === pageUrlStr) {
+      if (shouldIgnoreURL(sameSiteURLs, anchorURL)) {
         continue
       }
       if (anchorSite !== pageSite) {
         continue
       }
-      sameSiteLinkedUrls.add(processedAnchorUrlStr)
+      sameSiteURLs.add(anchorURL)
     } catch (e) {
-      console.error('getSameSitePageUrls')
+      console.error(`getSameSiteLinkedURLs: url ${pageURL}`)
       console.error(e)
     }
   }
 
-  return sameSiteLinkedUrls
+  return sameSiteURLs
 }
 
 module.exports = {
   closeAllPages,
   launchBrowser,
-  getPageContentForUrl,
-  getSameSitePageUrls
+  getPageContentForURL,
+  getSameSiteLinkedURLs
 }